backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
supabase==2.4.6
openai==1.30.1
feedparser==6.0.11
snscrape==0.7.0.20230622
firebase-admin==6.5.0
resend==2.1.0
python-multipart==0.0.9
pgvector==0.2.5
numpy==1.26.4
